feat: 시설검색 및 시설예약 구현

+++ src/main/java/com/lib/ailibrary/domain/room/RoomReserveRequest.java
@@ -13,13 +13,11 @@ public class RoomReserveRequest {
 
     private Long rezId;         // PK
     private int rezPeopleNum;   // 시설 사용인원
-
     @JsonFormat(pattern = "yyyy-MM-dd")
     private LocalDate rezDate;       // 예약 날짜
-
     private String[] rezTime;     // 예약 시간
     private int roomId;         // 시설번호
-    private int userStuId;   // 예약자 학번
+    private Long userStuId;   // 예약자 학번
     private String userName;    // 예약자 성명
 
 }
